backend/requirements.txt
# MedScribe Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
websockets==12.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.3
cryptography==42.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & NLP
anthropic==0.25.0
spacy==3.7.4

# PDF Generation
reportlab==4.2.0

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5

# Configuration
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0

# Utilities
python-dateutil==2.9.0
uuid6==2024.1.12

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0

# Rate Limiting
slowapi==0.1.9
